+++ PowerShell/JumpCloud Module/JumpCloud.psm1
@@ -1,7 +1,6 @@
+# Load all functions from public and private folders
 $Public = @( Get-ChildItem -Path "$PSScriptRoot/Public/*.ps1" -Recurse )
-
 $Private = @( Get-ChildItem -Path "$PSScriptRoot/Private/*.ps1" -Recurse)
-
 Foreach ($Import in @($Public + $Private))
 {
     Try
@@ -13,7 +12,10 @@ Foreach ($Import in @($Public + $Private))
         Write-Error -Message "Failed to import function $($Import.FullName): $_"
     }
 }
-# Function Aliases
+# Set default values for function parameters
+$PSDefaultParameterValues['Invoke-RestMethod:ContentType'] = 'application/json; charset=utf-8'
+$PSDefaultParameterValues['Invoke-WebRequest:ContentType'] = 'application/json; charset=utf-8'
+# Set function aliases
 Set-Alias -Name:('New-JCAssociation') -Value:('Add-JCAssociation')
-# Export Module Member
+# Export module member
 Export-ModuleMember -Function $Public.BaseName -Alias *