+++ PowerShell/JumpCloud Module/Public/RadiusServer/Add-JCRadiusReplyAttribute.ps1
@@ -37,6 +37,10 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
             $NumberOfAttributes = 2
             $VLAN = 11
         }
+        ElseIf ([System.String]::IsNullOrEmpty($NumberOfAttributes))
+        {
+            $NumberOfAttributes = 0
+        }
         $dict = New-Object System.Management.Automation.RuntimeDefinedParameterDictionary
 
         [int]$NewParams = 0
@@ -91,7 +95,7 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
     {
 
         Write-Verbose 'Verifying JCAPI Key'
-        if ($JCAPIKEY.length -ne 40) {Connect-JConline}
+        if ($JCAPIKEY.length -ne 40) { Connect-JCOnline }
 
         $hdrs = @{
 
@@ -115,7 +119,7 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
         $ResultsArray = @()
     }
 
-    process
+    Process
     {
 
         if ($GroupNameHash.containsKey($GroupName))
@@ -131,11 +135,11 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
 
             $ExistingAttributes = $GroupInfo | Select-Object -ExpandProperty attributes
         }
-        else { Throw "Group does not exist. Run 'Get-JCGroup -type User' to see a list of all your JumpCloud user groups."}
+        else { Throw "Group does not exist. Run 'Get-JCGroup -type User' to see a list of all your JumpCloud user groups." }
 
         $replyAttributes = New-Object System.Collections.ArrayList
 
-        $CurrentAttributes = Get-JCGroup -Type User -Name $GroupName | Select-Object @{Name = "RadiusAttributes"; Expression = {$_.attributes.radius.reply}} | Select-Object -ExpandProperty RadiusAttributes
+        $CurrentAttributes = Get-JCGroup -Type User -Name $GroupName | Select-Object @{Name = "RadiusAttributes"; Expression = { $_.attributes.radius.reply } } | Select-Object -ExpandProperty RadiusAttributes
 
         $RadiusCustomAttributesArrayList = New-Object System.Collections.ArrayList
 
@@ -179,7 +183,7 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
 
         }
 
-        $NewAttributesHash = @{}
+        $NewAttributesHash = @{ }
 
         foreach ($NewA in $NewAttributes)
         {
@@ -239,7 +243,7 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
 
         }
 
-        $CurrentAttributesHash = @{}
+        $CurrentAttributesHash = @{ }
 
         $VLANAttrHash = @{
             "Tunnel-Type"             = "Tunnel-Type"
@@ -253,7 +257,7 @@ The value specified for the ''-VLAN'' parameter is populated for the value of **
             {
                 $TagSplit = ($CurrentA.name -split ":")[0]
 
-                if (($VLANAttrHash).ContainsKey($TagSplit)) {Continue}
+                if (($VLANAttrHash).ContainsKey($TagSplit)) { Continue }
 
                 else
                 {