+++ PowerShell/JumpCloud Module/Public/Systems/Get-JCSystem.ps1
@@ -132,8 +132,8 @@ Function Get-JCSystem ()
         [Parameter(
             ValueFromPipelineByPropertyName,
             ParameterSetName = 'SearchFilter',
-            HelpMessage = 'A parameter that can filter on the property ''created''. This parameter if used creates two more dynamic parameters ''dateFilter'' and ''date''. See EXAMPLE 5 above for full syntax.')]
-        [ValidateSet('created')]
+            HelpMessage = 'A parameter that can filter on the property ''created'' or ''lastContact''. Only inactive systems will be returned when using the lastContact filter. This parameter if used creates two more dynamic parameters ''dateFilter'' and ''date''. See EXAMPLE 5 above for full syntax.')]
+        [ValidateSet('created', 'lastContact')]
         [String]$filterDateProperty,
 
         [Parameter(