+++ PowerShell/JumpCloud Module/Tests/Public/Systems/Get-JCSystem.tests.ps1
@@ -297,3 +297,16 @@ Describe -Tag:('JCSystem') "Get-JCSystem 1.4" {
     }
 
 }
+
+Describe -Tag:('JCSystem') "Get-JCSystem 1.15.2" {
+    It "Searches for a JumpCloud system using filterDateProperty lastContact and after" {
+        $NewestSystemlastContact = ([DateTime](Get-JCSystem -active $false -returnProperties lastContact | Where-Object lastContact -ne $null | Select-Object -Last 1 -ExpandProperty lastContact)).ToUniversalTime()
+        $NewestSystemlastContactVerify = ([DateTime](Get-JCSystem -filterDateProperty lastContact -dateFilter after -date $NewestSystemlastContact.addSeconds(-1) | Select-Object -Last 1 -ExpandProperty lastContact)).ToUniversalTime()
+        $NewestSystemlastContact | Should -Be $NewestSystemlastContactVerify
+    }
+    It "Searches for a JumpCloud system using filterDateProperty lastContact and before" {
+        $OldestSystemlastContact = ([DateTime](Get-JCSystem -active $false -returnProperties lastContact | Where-Object lastContact -ne $null | Select-Object -First 1 -ExpandProperty lastContact)).ToUniversalTime()
+        $OldestSystemlastContactVerify = ([DateTime](Get-JCSystem -filterDateProperty lastContact -dateFilter before -date $OldestSystemlastContact.addSeconds(1) | Select-Object -First 1 -ExpandProperty lastContact)).ToUniversalTime()
+        $OldestSystemlastContact | Should -Be $OldestSystemlastContactVerify
+    }
+}